% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Is.csv.R
\name{is.csv}
\alias{is.csv}
\title{Check if str is name of a CSV file or not}
\usage{
is.csv(str)
}
\arguments{
\item{a}{character vector including names of a file}
}
\value{
return true or false whether the name ended with '.csv' or not
}
\description{
Check if str is name of a CSV file or not
}
